requirements.txt
fastapi
uvicorn
passlib[bcrypt]
python-jose[cryptography]
pytest
flake8
python-multipart
